refactor(inbox): extract selection-handoff functions to inboxSelectionModel
Move pickReclassifyTarget, resolveReclassifyHandoff, and
resolveClassifiedGroupSelection to their own module. Re-exported from
InboxPage for backward-compatible test imports.
